REFACTOR: Implement structural updates and optimizations across multiple modules
FEAT: Implement methods to get the perfect position with the width height and lenght of a 3D Model, but need to be optimized. Otherwise, it's causing a FPS downfall

FEAT: Trying to implement background model, but it seems not loaded

+++ gui/include/graphical/Graphics.hpp
@@ -26,7 +26,7 @@
 namespace Gui {
     constexpr size_t WINDOW_WIDTH = 1200;
     constexpr size_t WINDOW_HEIGHT = 600;
-    constexpr size_t TILE_SIZE = 5;
+    constexpr float TILE_SIZE = 5.0;
 
     class Graphics {
         public:
@@ -41,7 +41,6 @@ namespace Gui {
                 SCOREBOARD,
             };
         public:
-            // Modifiez le constructeur pour prendre QueueManager au lieu des queues
             Graphics(std::shared_ptr<QueueManager> queueManager);
             ~Graphics();
 
@@ -58,6 +57,7 @@ namespace Gui {
         ////////////////////////////////////// GRAPHIC //////////////////////////////////////
             void init(void);
             void run(std::atomic<bool> &);
+
         private:
             void handleEvents(void);
             void update(void);
